feat:  컴포넌트에 로그인|회원가입 버튼 추가 - 컴포넌트 이름 변경 MainHeader -> Header - token 존재 시 회원관리 툴팁 아이콘 표시
