+++ frontend/src/App.js
@@ -1,20 +1,25 @@
 import React from "react";
 import { BrowserRouter, Route, Switch } from "react-router-dom";
-import "./api/api";
+
 import Footer from "./components/Footer/Footer";
-import MainHeader from "./components/MainHeader/MainHeader";
+import Header from "./components/Header/Header";
 import PrivateRoute from "./components/PrivateRoute/PrivateRoute";
+import ScrollToTop from "./components/ScrollToTop/ScrollToTop";
+
+import Recruits from "./pages/Recruits/Recruits";
 import ApplicantRegister from "./pages/ApplicantRegister/ApplicantRegister";
 import ApplicationRegister from "./pages/ApplicationRegister/ApplicationRegister";
 import Login from "./pages/Login/Login";
 import PasswordEdit from "./pages/PasswordEdit/PasswordEdit";
 import PasswordFind from "./pages/PasswordFind/PasswordFind";
 import PasswordFindResult from "./pages/PasswordFindResult/PasswordFindResult";
-import Recruits from "./pages/Recruits/Recruits";
+
 import RecruitmentProvider from "./provider/RecruitmentProvider";
 import TokenProvider from "./provider/TokenProvider";
-import ScrollToTop from "./components/ScrollToTop/ScrollToTop";
+import "./api/api";
+
 import PATH from "./constants/path";
+
 import "./App.css";
 
 const App = () => {
@@ -22,7 +27,7 @@ const App = () => {
     <TokenProvider>
       <RecruitmentProvider>
         <BrowserRouter>
-          <MainHeader />
+          <Header />
 
           <main className="main">
             <ScrollToTop>
