+++ frontend/src/components/Header/Header.module.css
@@ -1,0 +1,53 @@
+.container {
+  margin-bottom: 93px;
+}
+
+.header {
+  position: fixed;
+  top: 0;
+  width: 100%;
+  height: 5.75rem;
+  display: flex;
+  justify-content: center;
+  align-items: center;
+  background: var(--white);
+  border-bottom: 1px solid var(--gray-003);
+  z-index: 9999;
+}
+
+.content {
+  width: var(--pc-main-width);
+  display: flex;
+  justify-content: space-between;
+}
+
+.title {
+  font-size: 1.375rem;
+  display: flex;
+  align-items: center;
+  color: var(--black);
+}
+
+.logo {
+  width: 200px;
+  margin-right: 0.25rem;
+  transform: translateX(-6px);
+}
+
+.link-container {
+  display: flex;
+  align-items: center;
+  gap: 0.75rem;
+  font-size: 0.875rem;
+  font-weight: bold;
+}
+
+.bar {
+  width: 1px;
+  height: 1.25rem;
+  background-color: var(--gray-008);
+}
+
+.link-container a {
+  color: var(--gray-008);
+}
