+++ frontend/src/components/Header/Header.stories.js
@@ -1,0 +1,13 @@
+import React from "react";
+import Header from "./Header";
+
+export default {
+  title: "Header",
+  component: Header,
+};
+
+const Template = (args) => <Header {...args} />;
+
+export const Default = Template.bind({});
+
+Default.args = {};
