--- frontend/src/components/MainHeader/MainHeader.module.css
@@ -1,38 +1,0 @@
-.main-header {
-  margin-bottom: 93px;
-  user-select: none;
-}
-
-.header {
-  position: fixed;
-  top: 0;
-  display: flex;
-  justify-content: center;
-  align-items: center;
-  width: 100%;
-  height: 93px;
-  background: #fff;
-  box-shadow: 0px 4px 6px 0px rgba(12, 0, 46, 0.06);
-  z-index: 9999;
-}
-
-.nav {
-  display: flex;
-  align-items: center;
-  width: 900px;
-}
-
-.logo {
-  width: 200px;
-}
-
-.title {
-  font-size: 22px;
-  display: flex;
-  align-items: center;
-}
-
-a {
-  text-decoration: none;
-  color: #444;
-}
