--- frontend/src/components/MainHeader/MainHeader.stories.js
@@ -1,13 +1,0 @@
-import React from "react";
-import MainHeader from "./MainHeader";
-
-export default {
-  title: "MainHeader",
-  component: MainHeader,
-};
-
-const Template = (args) => <MainHeader {...args} />;
-
-export const Default = Template.bind({});
-
-Default.args = {};
